+++ pystac/__init__.py
@@ -67,7 +67,7 @@ from pystac.media_type import MediaType
 from pystac.rel_type import RelType
 from pystac.stac_io import StacIO
 from pystac.stac_object import STACObject, STACObjectType
-from pystac.link import Link, HIERARCHICAL_LINKS, HREF
+from pystac.link import Link, HIERARCHICAL_LINKS
 from pystac.catalog import Catalog, CatalogType
 from pystac.collection import (
     Collection,
@@ -81,9 +81,11 @@ from pystac.asset import Asset
 from pystac.item import Item
 from pystac.item_collection import ItemCollection
 from pystac.provider import ProviderRole, Provider
+from pystac.utils import HREF
 import pystac.validation
 
 import pystac.extensions.hooks
+import pystac.extensions.classification
 import pystac.extensions.datacube
 import pystac.extensions.eo
 import pystac.extensions.file
@@ -105,6 +107,7 @@ import pystac.extensions.view
 
 EXTENSION_HOOKS = pystac.extensions.hooks.RegisteredExtensionHooks(
     [
+        pystac.extensions.classification.CLASSIFICATION_EXTENSION_HOOKS,
         pystac.extensions.datacube.DATACUBE_EXTENSION_HOOKS,
         pystac.extensions.eo.EO_EXTENSION_HOOKS,
         pystac.extensions.file.FILE_EXTENSION_HOOKS,